+++ packages/vue/src/pressable/pressable.stories.vue
@@ -1,8 +1,0 @@
-<script setup lang="ts">
-import { Pressable } from '.'
-
-const handlePress = () => console.log('onPress')
-</script>
-<template>
-  <Pressable @press="handlePress">Pressable</Pressable>
-</template>